GoHome() finished for now
Added the function call to orient the robot for indexing.  When the bot
enters the indexing function, it will be stopped, facing the side wall,
and the grip line tracker will be positioned over the last dark strip.
Code to keep the robot travelling a set distance away from the wall
needs to be added (into move_left() and move_right())